+++ Template/config/weekhelper_config.php
@@ -35,6 +35,21 @@
 
     </div>
 
+    <br>
+    <br>
+
+    <div class="task-form-container">
+
+        <div class="task-form-main-column">
+            <?= $this->form->label(t('Sticky time box at the bottom'), 'time_box_enabled') ?>
+            <?= $this->form->checkbox('time_box_enabled', t('enabled'), 1, $time_box_enabled, '', [
+                'autofocus',
+                'tabindex="3"'
+            ]) ?>
+        </div>
+
+    </div>
+
 
 
     <div class="task-form-bottom">