+++ Template/time_box.php
@@ -1,0 +1,4 @@
+<div class="weekhelper-time-box-container">
+    <span id="currentDateTimeBox">
+    </span>
+</div>